requirements.txt
aiohttp>=3.8.3
async-timeout>=4.0.2
tenacity>=8.2.2
Flask>=2.2.5
pandas>=2.0
numpy>=1.24
scikit-learn>=1.2
joblib>=1.2
jinja2>=3.1
click>=8.1
python-dotenv>=1.0
pytest>=7.3
